feat(fe/be): - Draft of the FE without JSON schema #5 - Completed Autocomplete - Renamed some files
