+++ frontend/src/common/MainHeaderComponent.vue
